chore: move similar themes to one folder
