+++ styles/src/themes/atelier/atelier-heath-light.ts
@@ -1,31 +1,31 @@
 import chroma from "chroma-js"
 import { Meta } from "../common/colorScheme"
 import { colorRamp, createColorScheme } from "../common/ramps"
-import { metaCommon, name, buildSyntax, Variant } from "../common/atelier-common"
+import { metaCommon, name, buildSyntax, Variant } from "./common"
 
 const variant: Variant = {
     meta: {
-        name: `${name} Forest Light`,
+        name: `${name} Heath Light`,
         ...metaCommon,
-        url: "https://atelierbram.github.io/syntax-highlighting/atelier-schemes/forest/",
+        url: "https://atelierbram.github.io/syntax-highlighting/atelier-schemes/heath/",
     },
     colors: {
-        base00: "#f1efee",
-        base01: "#e6e2e0",
-        base02: "#a8a19f",
-        base03: "#9c9491",
-        base04: "#766e6b",
-        base05: "#68615e",
-        base06: "#2c2421",
-        base07: "#1b1918",
-        base08: "#f22c40",
-        base09: "#df5320",
-        base0A: "#c38418",
-        base0B: "#7b9726",
-        base0C: "#3d97b8",
-        base0D: "#407ee7",
-        base0E: "#6666ea",
-        base0F: "#c33ff3",
+        base00: "#f7f3f7",
+        base01: "#d8cad8",
+        base02: "#ab9bab",
+        base03: "#9e8f9e",
+        base04: "#776977",
+        base05: "#695d69",
+        base06: "#292329",
+        base07: "#1b181b",
+        base08: "#ca402b",
+        base09: "#a65926",
+        base0A: "#bb8a35",
+        base0B: "#918b3b",
+        base0C: "#159393",
+        base0D: "#516aec",
+        base0E: "#7b59c0",
+        base0F: "#cc33cc",
     },
 }
