+++ styles/src/themes/atelier/atelier-sulphurpool-light.ts
@@ -1,31 +1,31 @@
 import chroma from "chroma-js"
 import { Meta } from "../common/colorScheme"
 import { colorRamp, createColorScheme } from "../common/ramps"
-import { metaCommon, name, buildSyntax, Variant } from "../common/atelier-common"
+import { metaCommon, name, buildSyntax, Variant } from "./common"
 
 const variant: Variant = {
     meta: {
-        name: `${name} Heath Light`,
+        name: `${name} Sulphurpool Light`,
         ...metaCommon,
-        url: "https://atelierbram.github.io/syntax-highlighting/atelier-schemes/heath/",
+        url: "https://atelierbram.github.io/syntax-highlighting/atelier-schemes/sulphurpool/",
     },
     colors: {
-        base00: "#f7f3f7",
-        base01: "#d8cad8",
-        base02: "#ab9bab",
-        base03: "#9e8f9e",
-        base04: "#776977",
-        base05: "#695d69",
-        base06: "#292329",
-        base07: "#1b181b",
-        base08: "#ca402b",
-        base09: "#a65926",
-        base0A: "#bb8a35",
-        base0B: "#918b3b",
-        base0C: "#159393",
-        base0D: "#516aec",
-        base0E: "#7b59c0",
-        base0F: "#cc33cc",
+        base00: "#f5f7ff",
+        base01: "#dfe2f1",
+        base02: "#979db4",
+        base03: "#898ea4",
+        base04: "#6b7394",
+        base05: "#5e6687",
+        base06: "#293256",
+        base07: "#202746",
+        base08: "#c94922",
+        base09: "#c76b29",
+        base0A: "#c08b30",
+        base0B: "#ac9739",
+        base0C: "#22a2c9",
+        base0D: "#3d8fd1",
+        base0E: "#6679cc",
+        base0F: "#9c637a",
     },
 }
